refactor: NavLink component
Fix pageDetail prev and next link
