+++ src/pages/docs/index.js
@@ -3,8 +3,7 @@ import GatsbyLink from 'gatsby-link'
 import get from 'lodash/get'
 import styled from 'styled-components'
 
-import { Banner, Nav } from 'components/Doc'
-import Header from 'components/Header'
+import { Banner, Nav, Header } from 'components/Doc'
 import Layout from 'layout'
 
 const getChildLink = item => {
@@ -25,7 +24,7 @@ export default class Docs extends Component {
 
     return (
       <Layout>
-        <Header isBlankBg />
+        <Header/>
         <Wrapper>
           <Banner />
           <Nav />
